requirements.txt
absl-py==0.13.0
aiocontextvars==0.2.2
apex==0.1
cachetools==4.2.2
certifi==2021.5.30
chardet==3.0.4
charset-normalizer==2.0.4
colorama==0.4.4
contextvars==2.4
cycler==0.10.0
Cython==0.29.24
decorator==4.4.2
defusedxml==0.6.0
flatbuffers==2.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
graphviz==0.8.4
grpcio==1.40.0
idna==2.6
imageio==2.9.0
immutables==0.16
importlib-metadata==4.8.1
kiwisolver==1.3.1
loguru==0.5.3
Markdown==3.3.4
matplotlib==3.3.4
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
networkx==2.5
ninja==1.10.2
oauthlib==3.1.1
olefile==0.46
onnx==1.8.1
onnx-simplifier==0.3.5
onnxoptimizer==0.2.6
onnxruntime==1.8.0
opencv-python==4.5.3.56
Pillow==8.3.2
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
PyMySQL==1.0.2
pyparsing==2.4.7
python-dateutil==2.8.2
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.17.2
scipy==1.5.4
tabulate==0.8.9
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
test-generator==0.1.1
thop==0.0.31.post2005241907
tifffile==2020.9.3
torch==1.9.0
torchaudio==0.9.0
torchvision==0.10.0
tqdm==4.62.2
typing-extensions==3.7.4.3
urllib3==1.22
Werkzeug==2.0.1
win32-setctime==1.0.3
wincertstore==0.2
yolox==0.1.0
zipp==3.5.0
